Try a different approach
Instead of a complicated, bash based, multi-step workflow
I'm just going to rewrite this all to use Javascript
and be coordianted in that regard.

Also, introduces the concept of strategies, which map to
different filing dashboard vendors, so that it is easy
to add support for agencies that don't use Netfile or
one of the other big vendors.